Move DisabledStates to nested class inside TreeValuesCache
It's only used within AddAlertTemplate/CollectDisabledStates, so a
separate file under Cache/ is unnecessary. Also removes the standalone
DisabledStates.cs file.
